+++ packages/frontend/src/hooks/useSavedQuery.ts
@@ -43,11 +43,12 @@ const createSavedQuery = async (
 const duplicateSavedQuery = async (
     projectUuid: string,
     chartUuid: string,
+    data: { chartName: string; chartDesc: string },
 ): Promise<SavedChart> =>
     lightdashApi<SavedChart>({
         url: `/projects/${projectUuid}/saved?duplicateFrom=${chartUuid}`,
         method: 'POST',
-        body: undefined,
+        body: JSON.stringify(data),
     });
 
 export const deleteSavedQuery = async (id: string) =>
@@ -407,8 +408,16 @@ export const useDuplicateChartMutation = (
     const { projectUuid } = useParams<{ projectUuid: string }>();
     const queryClient = useQueryClient();
     const { showToastSuccess, showToastError } = useToaster();
-    return useMutation<SavedChart, ApiError, SavedChart['uuid']>(
-        (chartUuid) => duplicateSavedQuery(projectUuid, chartUuid),
+    return useMutation<
+        SavedChart,
+        ApiError,
+        Pick<SavedChart, 'uuid' | 'name' | 'description'>
+    >(
+        ({ uuid, name, description }) =>
+            duplicateSavedQuery(projectUuid, uuid, {
+                chartName: name,
+                chartDesc: description ?? '',
+            }),
         {
             mutationKey: ['saved_query_create', projectUuid],
             onSuccess: async (data) => {